P5 verifier fixes: durability goes double; range + associativity criticals
Adversarial verify (5 agents) found 3 reachable criticals the fixtures
missed: InteractionRange was 3.0 vs Python 3.5; efficiency_mult summed
bonuses in the wrong association (1.0+a+b vs 1.0+(a+b) — whole-point
damage differences on live title+enchant combos); and the fractional-
durability exactness map desynced against Repair/combat writers. Root
fix: EquipmentItem.DurabilityCurrent is now a DOUBLE exactly like
Python's float field — map deleted, every reader (effectiveness curve,
repair, defense, gates) exact by construction. Also: Fortune breaks
only on a successful roll (failed rolls keep scanning, consuming
draws), fallback loot collapsed to Python's EFFECTIVE 29-key dict
(alias literals overwrote canonical entries), TitleOrder cleared on
reload, regen-gate fields (time-since-damage, in-combat) wired for the
future update tick. Suite 89/89 green through the refactor.

+++ Game-1-Godot/src/Game1.Core/Data/EquipmentItem.cs
@@ -21,7 +21,9 @@ public sealed class EquipmentItem
     public required string Slot { get; init; }
     public (int Min, int Max) Damage { get; init; }
     public int Defense { get; init; }
-    public int DurabilityCurrent { get; set; } = 100;
+    // Python durability_current is a FLOAT (DEF-scaled fractional losses);
+    // double here keeps every read (effectiveness, repair, gates) exact.
+    public double DurabilityCurrent { get; set; } = 100;
     public int DurabilityMax { get; init; } = 100;
     public double AttackSpeed { get; init; } = 1.0;
     public double Efficiency { get; set; } = 1.0;
@@ -45,12 +47,12 @@ public sealed class EquipmentItem
     {
         if (DurabilityCurrent <= 0)
             return 0.5;
-        var durPct = (double)DurabilityCurrent / DurabilityMax;
+        var durPct = DurabilityCurrent / DurabilityMax;
         return durPct >= 0.5 ? 1.0 : 1.0 - (0.5 - durPct) * 0.5;
     }
 
     // :57-80
-    public int Repair(int? amount = null, double? percent = null)
+    public double Repair(int? amount = null, double? percent = null)
     {
         var old = DurabilityCurrent;
         if (amount is not null)
@@ -67,7 +69,7 @@ public sealed class EquipmentItem
     public string GetRepairUrgency()
     {
         if (DurabilityCurrent >= DurabilityMax) return "none";
-        var percent = (double)DurabilityCurrent / DurabilityMax;
+        var percent = DurabilityCurrent / DurabilityMax;
         if (percent >= 0.5) return "low";
         if (percent >= 0.2) return "medium";
         return percent > 0 ? "high" : "critical";
